requirements.txt
# Signals SaaS Backend Dependencies
# Investment-grade portfolio analysis

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP Client
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Financial Data
yfinance>=0.2.31

# Statistical Analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Text Matching
rapidfuzz>=3.5.0

# Visualization
matplotlib>=3.8.0

# AI Integration (optional, for memo generation)
openai>=1.3.0

# Production Server
gunicorn>=21.2.0

PyJWT>=2.8.0
cryptography>=41.0.0
